requirements.txt
customtkinter
mysql-connector-python
requests
cryptography
ping3